requirements.txt
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0
rich>=13.7.0